Add DTOs

+++ Models/ContaObjetivo.cs
@@ -7,15 +7,15 @@ namespace PoupaDevAPI.Models
 {
     public class ContaObjetivo : Entity<int>
     {
-        public decimal SaldoAtual { get; set; }
-        public decimal SaldoAnterior { get; set; }
+        public float SaldoAtual { get; set; }
+        public float SaldoAnterior { get; set; }
         public int ObjetivoFinanceiroId { get; set; }        
         public ObjetivoFinanceiro ObjetivoFinanceiro { get; set; }
         public bool EstaDeletado { get; set; }
         public List<OperacaoFinanceira> ListaOperacoes { get; set; }
 
 
-        public ContaObjetivo(decimal saldoAtual, decimal saldoAnterior, int objetivoFinanceiroId)
+        public ContaObjetivo(float saldoAtual, float saldoAnterior, int objetivoFinanceiroId)
         {
             SaldoAtual = saldoAtual;
             SaldoAnterior = saldoAnterior;
